mv eval_tuple and similar to pypeit.utils
