+++ pypeit/par/util.py
@@ -4,106 +4,17 @@ Utility functions for PypeIt parameter sets
 
 .. include:: ../include/links.rst
 """
-import ast
-import itertools
-from typing import Type
 
 from configobj import ConfigObj
 from IPython import embed
 
 from pypeit import PypeItError
+from pypeit import utils
 
 
 def _eval_ignore():
     """Provides a list of strings that should not be evaluated."""
     return [ 'open', 'file', 'dict', 'list', 'tuple' ]
-
-
-def ast_literal_eval(inp):
-    """
-    A wrapper for :func:`ast.literal_eval` that returns the input if it raises a
-    ValueError.
-    """
-    try:
-        return ast.literal_eval(inp)
-    except ValueError:
-        return inp
-
-
-def _eval_iter(inp:list[str], left:str, right:str, otype:Type) -> list:
-    """
-    Convenience function used to abstract the core functionality of
-    :func:`eval_tuple` and :func:`eval_list`.
-
-    Parameters
-    ----------
-    inp
-        Input list of strings
-    left
-        Left bracket delimeter
-    right
-        Right bracket delimeter
-    otype
-        Return type for the components of the list
-
-    Returns
-    -------
-        A list of objects with type ``otype`` as converted from the provided
-        strings.
-    """
-    grps = [
-        i for i in list(itertools.chain(*[s.split(right) for s in ','.join(inp).split(left)]))
-        if len(i) > 1
-    ]
-    return list(otype(map(ast_literal_eval, g.split(','))) for g in grps)
-
-
-def eval_tuple(inp:list[str]) -> list[tuple]:
-    """
-    Evaluate the input to one or more tuples.
-
-    This allows conversion of one or more tuples provided to a configuration
-    parameters.
-
-    .. warning::
-
-        - Currently can only handle tuples with integers, floats, or strings!
-
-    Parameters
-    ----------
-    inp 
-        A list of strings that are converted into a list of tuples.  The
-        parentheses must be within the list of elements.
-
-    Returns
-    -------
-        A list of tuples with the converted elements.
-    """
-    return _eval_iter(inp, '(', ')', tuple)
-
-
-def eval_list(inp:list[str]) -> list[list]:
-    """
-    Evaluate the input to one or more lists.
-
-    This allows conversion of one or more lists provided to a configuration
-    parameters.
-
-    .. warning::
-
-        - Currently can only handle lists with integers, floats, or strings!
-
-    Parameters
-    ----------
-    inp 
-        A list of strings that are converted into a list of lists.  The square
-        brackets must be within the list of elements.
-
-    Returns
-    -------
-        A list of lists with the converted elements.
-    """
-    return _eval_iter(inp, '[', ']', list)
 
 
 def recursive_dict_evaluate(d):
@@ -150,7 +61,7 @@ def recursive_dict_evaluate(d):
         if isinstance(d[k], list) and any(['(' in e for e in d[k]]):
             # NOTE: This enables syntax for constructing one or more tuples.
             try:
-                d[k] = eval_tuple(d[k])
+                d[k] = utils.eval_tuple(d[k])
             except (PypeItError, SyntaxError) as e:
                 # The tuple evaluation failed.  Assume that this can be handled
                 # later in the code and leave the dictionary element unaltered.
